[REFACTOR] 메소드 또는 API 단위로 테스트 분리
- 테스트 파일의 크기가 커져서 각 메소드 별 테스트 파악이 안됨
- 메소드는 메소드의 이름으로 파일 이름 구분
- API는 api 경로로 파일 이름 구분
- 파일 이름을 구분함으로써 찾고자 하는 테스트를 쉽게 찾을 수 있음
