+++ server/test/router/board/BoardRouter.putBoardByBoardId.test.js
@@ -1,0 +1,51 @@
+import request from 'supertest';
+import { getRepository } from 'typeorm';
+import { Application } from '../../../src/Application';
+import { JwtUtil } from '../../../src/common/util/JwtUtil';
+import { Board } from '../../../src/model/Board';
+import { User } from '../../../src/model/User';
+import { TransactionRollbackExecutor } from '../../TransactionRollbackExecutor';
+
+describe('PUT /api/board/:boardId', () => {
+    const app = new Application();
+    let jwtUtil = null;
+
+    beforeAll(async () => {
+        await app.initialize();
+        jwtUtil = JwtUtil.getInstance();
+    });
+
+    afterAll(async (done) => {
+        await app.close();
+        done();
+    });
+
+    test('PUT /api/board/:boardId가 정상적으로 호출되었을 때(권한o, 보드o), 204 리턴한다.', async () => {
+        await TransactionRollbackExecutor.rollback(async () => {
+            // given
+            const user = { name: 'dhoooon', socialId: 'naver-1', profileImageUrl: 'dhoon-amg' };
+            const userRepository = getRepository(User);
+            const createUser = userRepository.create(user);
+            const createdUser = await userRepository.save(createUser);
+
+            const boardRepository = getRepository(Board);
+            const board = { title: 'board', creator: createdUser.id, color: '#ffffff' };
+            const createBoard = boardRepository.create(board);
+            await boardRepository.save(createBoard);
+            const token = await jwtUtil.generateAccessToken({
+                userId: createdUser.id,
+                username: createdUser.name,
+            });
+            const boardId = createBoard.id;
+
+            // when
+            const response = await request(app.httpServer).put(`/api/board/${boardId}`).set({
+                Authorization: token,
+                'Content-Type': 'application/json',
+            });
+
+            // then
+            expect(response.status).toEqual(204);
+        });
+    });
+});
