+++ server/test/router/comment/CommentRouter.deleteCommentByCommentId.test.js
@@ -11,7 +11,7 @@ import { List } from '../../../src/model/List';
 import { User } from '../../../src/model/User';
 import { TransactionRollbackExecutor } from '../../TransactionRollbackExecutor';
 
-describe('PATCH /api/comment/{commentId}', () => {
+describe('DELETE /api/comment/:commentId', () => {
     const app = new Application();
     let jwtUtil = null;
 
@@ -25,7 +25,7 @@ describe('PATCH /api/comment/{commentId}', () => {
         done();
     });
 
-    test('존재하지 않는 댓글을 수정할 때 404 반환', async () => {
+    test('존재하지 않는 댓글을 삭제할 때 404 반환', async () => {
         await TransactionRollbackExecutor.rollback(async () => {
             // given
             const em = getEntityManagerOrTransactionManager('default');
@@ -38,11 +38,9 @@ describe('PATCH /api/comment/{commentId}', () => {
             const accessToken = await jwtUtil.generateAccessToken({ userId: user0.id });
             // when
             const response = await agent(app.httpServer)
-                .patch(`/api/comment/1`)
+                .delete(`/api/comment/1`)
                 .set('Authorization', accessToken)
-                .send({
-                    content: 'edited content',
-                });
+                .send();
             // then
             expect(response.status).toEqual(404);
         });
@@ -59,19 +57,17 @@ describe('PATCH /api/comment/{commentId}', () => {
             });
             await em.save(user0);
             const accessToken = await jwtUtil.generateAccessToken({ userId: user0.id });
-
             // when
             const response = await agent(app.httpServer)
-                .patch(`/api/comment/dd`)
+                .delete(`/api/comment/dd`)
                 .set('Authorization', accessToken)
                 .send();
-
             // then
             expect(response.status).toEqual(400);
         });
     });
 
-    test('본인 댓글이 아닌 댓글을 수정할 때 403 반환', async () => {
+    test('본인 댓글이 아닌 댓글을 삭제할 때 403 반환', async () => {
         await TransactionRollbackExecutor.rollback(async () => {
             // given
             const em = getEntityManagerOrTransactionManager('default');
@@ -123,15 +119,12 @@ describe('PATCH /api/comment/{commentId}', () => {
             await em.save(comment0);
 
             const accessToken = await jwtUtil.generateAccessToken({ userId: user0.id });
-            const expectedContent = 'edited content';
 
             // when
             const response = await agent(app.httpServer)
-                .patch(`/api/comment/${comment0.id}`)
+                .delete(`/api/comment/${comment0.id}`)
                 .set('Authorization', accessToken)
-                .send({
-                    content: expectedContent,
-                });
+                .send();
 
             // then
             expect(response.status).toEqual(403);
@@ -183,13 +176,12 @@ describe('PATCH /api/comment/{commentId}', () => {
             await em.save(comment0);
 
             const accessToken = await jwtUtil.generateAccessToken({ userId: user0.id });
-            const expectedContent = 'edited content';
 
             // when
             const response = await agent(app.httpServer)
-                .patch(`/api/comment/${comment0.id}`)
+                .delete(`/api/comment/${comment0.id}`)
                 .set('Authorization', accessToken)
-                .send({ content: expectedContent });
+                .send();
 
             // then
             expect(response.status).toEqual(204);
