+++ server/test/router/user/UserRotuer.getUserMe.test.js
@@ -1,12 +1,11 @@
-import request, { agent } from 'supertest';
+import { agent } from 'supertest';
 import { getEntityManagerOrTransactionManager } from 'typeorm-transactional-cls-hooked';
-import { getRepository } from 'typeorm';
-import { Application } from '../../src/Application';
-import { JwtUtil } from '../../src/common/util/JwtUtil';
-import { User } from '../../src/model/User';
-import { TransactionRollbackExecutor } from '../TransactionRollbackExecutor';
+import { Application } from '../../../src/Application';
+import { JwtUtil } from '../../../src/common/util/JwtUtil';
+import { TransactionRollbackExecutor } from '../../TransactionRollbackExecutor';
+import { User } from '../../../src/model/User';
 
-describe('UserRouter Test', () => {
+describe('GET /api/user/me', () => {
     const app = new Application();
     let jwtUtil = null;
 
@@ -48,29 +47,6 @@ describe('UserRouter Test', () => {
                 name: user0.name,
                 profileImageUrl: user0.profileImageUrl,
             });
-        });
-    });
-
-    test('GET /api/user?username= 유저 검색 라우터 정상호출 시 상태코드 200 반환', async () => {
-        await TransactionRollbackExecutor.rollback(async () => {
-            // given
-            const user = { name: 'dhoon', socialId: 'naver', profileImageUrl: 'image' };
-            const userRepository = getRepository(User);
-            const createUser = userRepository.create(user);
-            const createdUser = await userRepository.save(createUser);
-            const token = await jwtUtil.generateAccessToken({
-                userId: createdUser.id,
-                username: createdUser.name,
-            });
-
-            // when
-            const response = await request(app.httpServer).get('/api/user?username=dhoon').set({
-                Authorization: token,
-                'Content-Type': 'application/json',
-            });
-
-            // then
-            expect(response.status).toEqual(200);
         });
     });
 });
