+++ server/test/router/user/UserRouter.getUserByUsername.test.js
@@ -1,0 +1,44 @@
+import request from 'supertest';
+import { getRepository } from 'typeorm';
+import { Application } from '../../../src/Application';
+import { JwtUtil } from '../../../src/common/util/JwtUtil';
+import { TransactionRollbackExecutor } from '../../TransactionRollbackExecutor';
+import { User } from '../../../src/model/User';
+
+describe('GET /api/user?username=', () => {
+    const app = new Application();
+    let jwtUtil = null;
+
+    beforeAll(async () => {
+        await app.initialize();
+        jwtUtil = JwtUtil.getInstance();
+    });
+
+    afterAll(async (done) => {
+        await app.close();
+        done();
+    });
+
+    test('GET /api/user?username= 유저 검색 라우터 정상호출 시 상태코드 200 반환', async () => {
+        await TransactionRollbackExecutor.rollback(async () => {
+            // given
+            const user = { name: 'dhoon', socialId: 'naver', profileImageUrl: 'image' };
+            const userRepository = getRepository(User);
+            const createUser = userRepository.create(user);
+            const createdUser = await userRepository.save(createUser);
+            const token = await jwtUtil.generateAccessToken({
+                userId: createdUser.id,
+                username: createdUser.name,
+            });
+
+            // when
+            const response = await request(app.httpServer).get('/api/user?username=dhoon').set({
+                Authorization: token,
+                'Content-Type': 'application/json',
+            });
+
+            // then
+            expect(response.status).toEqual(200);
+        });
+    });
+});
